requirements.txt
anthropic>=0.42.0
psycopg2-binary>=2.9.9
streamlit>=1.40.0
python-dotenv>=1.0.0
exa-py>=1.0.0
requests>=2.31.0
